refactor store to remove require cycle
